requirements.txt
pygtail==0.7.0
hiredis==0.2.0
kafka-python==1.3.2
PyYAML==3.12